add more blitter code

+++ Sources/Blitter.h
@@ -122,6 +122,6 @@ private:
    // line drawing variable
    unsigned int line_x_;
    unsigned int line_y_;
-   unsigned int x_mod_;
+   unsigned short x_mod_;
    short remain_;
 };
